split up helix into more modules
